# api_documentation.yaml
# api_documentation.yaml

swagger: '2.0'
info:
  title: BigQuery Client API
  description: An API for interacting with Google BigQuery using Flask.
  version: 1.0.0

paths:
  /connect_to_project:
    get:
      summary: Connect to a BigQuery project.
      parameters:
        - name: project_id
          in: query
          type: string
          required: true
          description: The BigQuery project ID.
      responses:
        '200':
          description: Successful connection to the BigQuery project.

  /connect_to_dataset:
    get:
      summary: Connect to a BigQuery dataset.
      parameters:
        - name: dataset_id
          in: query
          type: string
          required: true
          description: The BigQuery dataset ID.
      responses:
        '200':
          description: Successful connection to the BigQuery dataset.

  /connect_to_table:
    get:
      summary: Connect to a BigQuery table.
      parameters:
        - name: table_id
          in: query
          type: string
          required: true
          description: The BigQuery table ID.
      responses:
        '200':
          description: Successful connection to the BigQuery table.

  /get_table_data:
    get:
      summary: Get data from a specified BigQuery table.
      parameters:
        - name: limit
          in: query
          type: integer
          required: false
          default: 10
          description: The maximum number of rows to retrieve.
      responses:
        '200':
          description: A list of records from the specified BigQuery table.

  /filter_table_data:
    get:
      summary: Filter data from a table based on dynamic filters.
      parameters:
        - name: limit
          in: query
          type: integer
          required: false
          description: The maximum number of rows to retrieve. If not specified, all rows are returned.
        - name: filters
          in: query
          type: object
          required: false
          description: Dynamic filters to apply to the table data.
      responses:
        '200':
          description: Successful response with filtered data.
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates the success of the request.
                example: true
              data:
                type: array
                description: List of records from the filtered table.
                items:
                  type: object  # Adjust based on the structure of your records
                  properties:
                    # Define properties based on your record structure

        '400':
          description: Invalid request. Check the error message for details.
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates the success of the request.
                example: false
              error:
                type: string
                description: Details about the error.
                example: "Invalid filter: invalid_column"

  /filter_row_data:
    get:
      summary: Filter rows based on a search value.
      parameters:
        - name: search_value
          in: query
          type: string
          required: true
          description: The value to search for in all rows.
      responses:
        '200':
          description: Successful response with matching rows.
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates the success of the request.
                example: true
              data:
                type: array
                description: List of records matching the search value.
                items:
                  type: object  # Adjust based on the structure of your records
                  properties:
                    # Define properties based on your record structure
        '400':
          description: Bad Request. The request is malformed or contains invalid parameters.
        '404':
          description: Not Found. The requested resource or endpoint does not exist.  